+++ trunk/python/casino/platinumslots.py
@@ -63,7 +63,7 @@ if (activator.PayAmount(cost*50)):#platinumcoin
             pay = slotmajor[slotlist.index(item)]
          else:
             break
-         payoff = pot*pay
+         payoff = int(pot*pay)
          Slots.payoff(payoff)
          id = Crossfire.CreateObjectByName(cointype)
          id.Quantity = payoff
